requirements.txt
streamlit
pandas
numpy
google-generativeai
sentence-transformers
torch